Optimizes PostExcerpt using React.memo
This addresses a bug where every PostExcerpt was being
re-rendered every time one of the Posts had one of its
reaction buttons clicked. This isn't a necessary re-render,
but PostsList was receiving a different posts obj when
one of the posts changed (via a changed reaction).

The solution is to memo() PostExcerpt so that it only
renders when its specific props change.